Primary engine is ready

+++ api/database.sql
@@ -28,7 +28,7 @@ CREATE TABLE url_index (
   is_indexed                         TINYINT(4) NOT NULL DEFAULT 0,
   total_words                        INT(11)    NOT NULL DEFAULT 0,
   time_elapsed_to_first_index_in_sec INT(11)             DEFAULT NULL,
-  last_indexed_at                    INT(11)             DEFAULT NULL,
+  last_indexed_at                    BIGINT                 DEFAULT NULL,
   created_at                         TIMESTAMP  NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id)
 );
